+++ drift_dev/CHANGELOG.md
@@ -3,6 +3,7 @@
 
 - Support `MAPPED BY` for individual columns in queries or in views defined with SQL.
 - Consistently interpret `CAST (x AS DATETIME)` and `CAST(x AS TEXT)` in drift files.
+- Support a `CAST` to an enum type in drift types.
 
 ## 2.4.1